+++ src/sdr/_filter/__init__.py
@@ -2,6 +2,7 @@
 A subpackage for various LTI filters.
 """
 from ._applications import *
+from ._delay import *
 from ._design_fir import *
 from ._design_multirate import *
 from ._fir import *
